add filtering for receipts

+++ app/assets/javascripts/general.coffee
@@ -1,7 +1,7 @@
 $(document).on 'turbolinks:load', ->
+
   $("body").on 'click', '.batch-destroy', (e) ->
     e.preventDefault()
-    return unless confirm('Are you sure?')
 
     url = $(@).prop('href')
     ids = []
@@ -10,10 +10,14 @@ $(document).on 'turbolinks:load', ->
       if $(@).prop('checked')
         ids.push $(@).data('id')
 
-    $.ajax {
-      url: url,
-      type: 'DELETE',
-      data: {ids: ids},
-      success: (result) ->
-        console.log('done')
-    }
+    if ids.length == 0
+      alert('No resources selected')
+    else
+      if confirm('Are you sure you want delete ' + ids.length + ' records?')
+        $.ajax {
+          url: url,
+          type: 'DELETE',
+          data: {ids: ids},
+          success: (result) ->
+            console.log('done')
+        }
